moved all logic to ToDoContainer
